Update library-management-system
For our CS221L semester project, we decided to implement a library management system using linked lists and queues. The information about each book is stored in the node of the linked list; including book name, book author, book ID, borrowed (integer which stores 0 if book hasn’t been borrowed and the registration number of the person who borrowed it, if it has been borrowed), and a queue which acts as a waiting list and stores the registration number of the next in line for the book if book has already been borrowed. As soon as a book is returned, the registration number of the next person in line is stored in the variable borrowed.